housekeeping
Move remaining relevant from `service` namespace to `svc` namespace
