+++ proto/CMakeLists.txt
@@ -5,6 +5,12 @@ cmake_path(SET access_grpc_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/access.grpc.
 cmake_path(SET access_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/access.pb.h)
 cmake_path(SET access_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/access.pb.cc)
 
+cmake_path(SET collections_proto ${CMAKE_CURRENT_SOURCE_DIR}/gk/v1/collections.proto)
+cmake_path(SET collections_grpc_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/collections.grpc.pb.h)
+cmake_path(SET collections_grpc_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/collections.grpc.pb.cc)
+cmake_path(SET collections_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/collections.pb.h)
+cmake_path(SET collections_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/collections.pb.cc)
+
 cmake_path(SET common_proto ${CMAKE_CURRENT_SOURCE_DIR}/gk/v1/common.proto)
 cmake_path(SET common_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/common.pb.h)
 cmake_path(SET common_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/common.pb.cc)
@@ -14,12 +20,6 @@ cmake_path(SET events_grpc_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/events.grpc.
 cmake_path(SET events_grpc_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/events.grpc.pb.cc)
 cmake_path(SET events_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/events.pb.h)
 cmake_path(SET events_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/events.pb.cc)
-
-cmake_path(SET gatekeeper_proto ${CMAKE_CURRENT_SOURCE_DIR}/gk/v1/gatekeeper.proto)
-cmake_path(SET gatekeeper_grpc_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/gatekeeper.grpc.pb.h)
-cmake_path(SET gatekeeper_grpc_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/gatekeeper.grpc.pb.cc)
-cmake_path(SET gatekeeper_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/gatekeeper.pb.h)
-cmake_path(SET gatekeeper_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/gatekeeper.pb.cc)
 
 cmake_path(SET identities_proto ${CMAKE_CURRENT_SOURCE_DIR}/gk/v1/identities.proto)
 cmake_path(SET identities_grpc_header ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/identities.grpc.pb.h)
@@ -41,9 +41,9 @@ cmake_path(SET roles_source ${CMAKE_CURRENT_BINARY_DIR}/gk/v1/roles.pb.cc)
 
 set(protos
 	${access_proto}
+	${collections_proto}
 	${common_proto}
 	${events_proto}
-	${gatekeeper_proto}
 	${identities_proto}
 	${rbac_proto}
 	${roles_proto}
@@ -51,10 +51,9 @@ set(protos
 
 set(headers
 	${access_header} ${access_grpc_header}
+	${collections_header} ${collections_grpc_header}
 	${common_header}
 	${events_header} ${events_grpc_header}
-	${gatekeeper_grpc_header}
-	${gatekeeper_header}
 	${identities_header} ${identities_grpc_header}
 	${rbac_header} ${rbac_grpc_header}
 	${roles_header} ${roles_grpc_header}
@@ -62,10 +61,9 @@ set(headers
 
 set(sources
 	${access_source} ${access_grpc_source}
+	${collections_source} ${collections_grpc_source}
 	${common_source}
 	${events_source} ${events_grpc_source}
-	${gatekeeper_grpc_source}
-	${gatekeeper_source}
 	${identities_source} ${identities_grpc_source}
 	${rbac_source} ${rbac_grpc_source}
 	${roles_source} ${roles_grpc_source}
