+++ src/svc/CMakeLists.txt
@@ -1,9 +1,12 @@
 add_library(svc
 	access.cpp access.h
+	collections.cpp collections.h
 	events.cpp events.h
 	identities.cpp identities.h
 	rbac.cpp rbac.h
 	roles.cpp roles.h
+	svc.h
+	interceptors/logger.cpp interceptors/logger.h
 )
 
 target_link_libraries(svc
@@ -12,6 +15,7 @@ target_link_libraries(svc
 	PRIVATE
 		${PROJECT_NAME}::datastore
 		${PROJECT_NAME}::err
+		${PROJECT_NAME}::logger
 )
 
 if (GATEKEEPER_ENABLE_COVERAGE)
@@ -28,6 +32,7 @@ endif()
 if (GATEKEEPER_ENABLE_TESTING)
 	add_executable(svc_tests
 		access_test.cpp
+		collections_test.cpp
 		events_test.cpp
 		identities_test.cpp
 		rbac_test.cpp
