+++ src/svc/interceptors/logger.cpp
@@ -2,7 +2,7 @@
 
 #include "logger/logger.h"
 
-namespace service {
+namespace svc {
 namespace interceptors {
 Logger::Logger(grpc::experimental::ServerRpcInfo *rpc) :
 	_rpc(rpc), _start(std::chrono::high_resolution_clock::now()) {}
@@ -11,11 +11,11 @@ Logger::~Logger() {
 	std::chrono::duration<double, std::milli> diff =
 		std::chrono::high_resolution_clock::now() - _start;
 
-	logger::info("service", "latency_ms", diff.count(), "method", _rpc->method());
+	logger::info("svc", "latency_ms", diff.count(), "method", _rpc->method());
 }
 
 void Logger::Intercept(grpc::experimental::InterceptorBatchMethods *methods) {
 	methods->Proceed();
 }
 } // namespace interceptors
-} // namespace service
+} // namespace svc
