+++ src/svc/svc.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "access.h"
+#include "collections.h"
+#include "events.h"
+#include "identities.h"
+#include "rbac.h"
+#include "roles.h"
